+++ x/perp/amm/genesis_test.go
@@ -19,9 +19,9 @@ func TestGenesis(t *testing.T) {
 	markets := []types.Market{
 		{
 			Pair:         asset.MustNewPair("BTC:NUSD"),
-			BaseReserve:  sdk.NewDec(1 * common.TO_MICRO),      // 1
-			QuoteReserve: sdk.NewDec(30_000 * common.TO_MICRO), // 30,000
-			SqrtDepth:    common.MustSqrtDec(sdk.NewDec(30_000 * common.TO_MICRO * common.TO_MICRO)),
+			BaseReserve:  sdk.NewDec(1 * common.TO_MICRO), // 1
+			QuoteReserve: sdk.NewDec(1 * common.TO_MICRO), // 30,000
+			SqrtDepth:    common.MustSqrtDec(sdk.NewDec(common.TO_MICRO * common.TO_MICRO)),
 			Config: types.MarketConfig{
 				TradeLimitRatio:        sdk.MustNewDecFromStr("0.88"),
 				FluctuationLimitRatio:  sdk.MustNewDecFromStr("0.20"),
@@ -30,13 +30,13 @@ func TestGenesis(t *testing.T) {
 				MaxLeverage:            sdk.MustNewDecFromStr("15"),
 			},
 			Bias:          sdk.NewDec(1 * common.TO_MICRO),
-			PegMultiplier: sdk.OneDec(),
+			PegMultiplier: sdk.NewDec(30_000),
 		},
 		{
 			Pair:         asset.MustNewPair("ETH:NUSD"),
-			BaseReserve:  sdk.NewDec(2 * common.TO_MICRO),      // 2
-			QuoteReserve: sdk.NewDec(60_000 * common.TO_MICRO), // 60,000
-			SqrtDepth:    common.MustSqrtDec(sdk.NewDec(2 * 60_000 * common.TO_MICRO * common.TO_MICRO)),
+			BaseReserve:  sdk.NewDec(2 * common.TO_MICRO), // 2
+			QuoteReserve: sdk.NewDec(2 * common.TO_MICRO), // 60,000
+			SqrtDepth:    common.MustSqrtDec(sdk.NewDec(2 * 2 * common.TO_MICRO * common.TO_MICRO)),
 			Config: types.MarketConfig{
 				TradeLimitRatio:        sdk.MustNewDecFromStr("0.77"),
 				FluctuationLimitRatio:  sdk.MustNewDecFromStr("0.30"),
@@ -45,7 +45,7 @@ func TestGenesis(t *testing.T) {
 				MaxLeverage:            sdk.MustNewDecFromStr("15"),
 			},
 			Bias:          sdk.NewDec(0),
-			PegMultiplier: sdk.MustNewDecFromStr("0.2"),
+			PegMultiplier: sdk.MustNewDecFromStr("60000"),
 		},
 	}